% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fire_locations.R
\name{plot_fire_locations}
\alias{plot_fire_locations}
\title{Plot Fire Locations}
\usage{
plot_fire_locations()
}
\value{
A ggplot object.
}
\description{
This function reads fire data from a CSV file and plots fire locations on a map.
}
