% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_predictions.R
\name{predict_fire}
\alias{predict_fire}
\title{Predict Fire Function}
\usage{
predict_fire(input_data)
}
\arguments{
\item{input_data}{A data frame containing the input data.}
}
\value{
A vector of predictions.
}
\description{
This function preprocesses input data and uses a pre-trained XGBoost model to make predictions.
}
